Move Loggers into its own file.
